Move has_z, remove normalize, finish get_x and get_y
