+++ pygeos/__init__.py
@@ -1,4 +1,6 @@
 from .ufuncs import GEOSException  # NOQA
+from .ufuncs import Geometry  # NOQA
+from .ufuncs import Empty  # NOQA
 from .geometry import *
 from .creation import *
 from .constructive import *
